moved handledelete to Posts.js
