+++ src/components/Posts/PostCard.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { useContext } from 'react';
 import { UserContext } from '../../context/UserContext';
-import { deletePost } from '../../services/posts';
 import { usePosts } from '../../hooks/usePosts';
 
 
@@ -11,17 +10,7 @@ export default function PostCard({ title, description, user_id, id }) {
   const owner = user.id === user_id;
 
   const { posts, setPosts } = usePosts();
-  const handleDelete = async () => {
-    try {
 
-      await deletePost(id);
-      const result = posts.filter(post => post.id !== id);
-      setPosts(result);
-    } catch (e) {
-      //eslint-disable-next-line no-console
-      console.error(e.message);
-    }
-  }; 
   return (
     <div className="post">
       <h3>{title}</h3>
