+++ src/components/Posts/Posts.js
@@ -1,13 +1,25 @@
 import React from 'react';
 import { usePosts } from '../../hooks/usePosts';
 import PostCard from './PostCard';
-
+import { deletePost } from '../../services/posts';
 
 export default function Posts() {
-  const { loading, error, posts } = usePosts();
+  const { loading, error, posts, setPosts } = usePosts();
   if (loading) return <h1>Loading</h1>;
   if (error) return <h1>{error}</h1>;
 
+  const handleDelete = async () => {
+    try {
+
+      await deletePost(id);
+      const result = posts.filter(post => post.id !== id);
+      await setPosts(result);
+
+    } catch (e) {
+      //eslint-disable-next-line no-console
+      console.error(e.message);
+    }
+  }; 
   return (
     <div>
       {posts.map((post) => (
